Return last on or last off value when Balance on_for and on_off are called
Without this change, we can't easily check for how long we were off balance for
when we regain balance. Since the current on and off value can be checked via
existing is_on and is_off functions already, there's no need to return False in
on_for and off_for.