Fix bug with duplicate writes to connected clients
Introduced in b9e9b95. When receivers.write(...) is called, we write to each of
the receivers, which in turn write to every transport in
self.factory.transports. However, self.factory.transports is the same between
each receiver, causing each transport to be written to len(receivers) number of
times instead of once.